Debug Mode
Created a debug mode that can be toggled in play mode or in a build using the '~' key. Will let us have a debug mode as we fully develop the project.

For now debug lets you turn on PC controls to make testing while developing easier, and also allows you to reset the game at any time.

Also cleaned up some code, object names, and changed the score system to a lives system for how our games actually gonna be.